# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Shopify API",
        "description": "RESTful API for managing Shopify products and variants with GraphQL integration",
        "contact": {
            "name": "API Support",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/api/v1/search/products": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Advanced product search",
                "description": "Search products with multiple filters and sorting options",
                "operationId": "7610860f1e4fb6b72f7f5a97d5815444",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query for title, description, vendor, SKU",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by product status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "draft",
                                "archived"
                            ]
                        }
                    },
                    {
                        "name": "vendor",
                        "in": "query",
                        "description": "Filter by vendor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_type",
                        "in": "query",
                        "description": "Filter by product type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "decimal"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "decimal"
                        }
                    },
                    {
                        "name": "has_shopify_id",
                        "in": "query",
                        "description": "Filter by Shopify sync status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "title",
                                "price",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Include related resources",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "variants",
                                "images",
                                "variants,images"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductWithRelations"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "count": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                },
                                                "filters_applied": {
                                                    "type": "object"
                                                },
                                                "search_time_ms": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/search/filters": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get available filters",
                "description": "Get available filter options and their counts",
                "operationId": "b91ff1a2fff51aa8ef8492a02cb04f18",
                "parameters": [
                    {
                        "name": "filter_type",
                        "in": "query",
                        "description": "Type of filters to return",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "status",
                                "vendor",
                                "product_type",
                                "all"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available filters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filters": {
                                            "type": "object",
                                            "example": {
                                                "status": {
                                                    "active": 150,
                                                    "draft": 25,
                                                    "archived": 10
                                                },
                                                "vendors": {
                                                    "BrandA": 50,
                                                    "BrandB": 30,
                                                    "BrandC": 20
                                                },
                                                "product_types": {
                                                    "Clothing": 80,
                                                    "Accessories": 40,
                                                    "Electronics": 25
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductWithRelations": {
                "title": "Product with Relations",
                "description": "Product model with variants and images loaded",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Product"
                    },
                    {
                        "properties": {
                            "variants": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Variant"
                                }
                            },
                            "images": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ProductImage"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token authentication. Use the token returned from /api/v1/auth/login endpoint in the format: Bearer {your_token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}